refactor: split sections into lazy components
